refactor: split srs_singbox.go into srs_parse, srs_skip, srs_cidr (711→~250 lines)
